+++ Source/AbyssDiverUnderWorld/Gimmic/Spawn/SpawnPoint/Monster/MonsterSpawnPoint.cpp
@@ -1,0 +1,2 @@
+#include "Gimmic/Spawn/SpawnPoint/Monster/MonsterSpawnPoint.h"
+
