+++ Source/AbyssDiverUnderWorld/Gimmic/Spawn/SpawnPoint/Ore/OreSpawnPoint.cpp
@@ -1,0 +1,2 @@
+#include "Gimmic/Spawn/SpawnPoint/Ore/OreSpawnPoint.h"
+
