+++ Source/AbyssDiverUnderWorld/Gimmic/Spawn/SpawnPoint/SpawnPoint.cpp
@@ -1,0 +1,2 @@
+#include "Gimmic/Spawn/SpawnPoint/SpawnPoint.h"
+
