--- Source/AbyssDiverUnderWorld/Gimmic/SpawnPoint/Monster/MonsterSpawnPoint.cpp
@@ -1,2 +1,0 @@
-#include "Gimmic/SpawnPoint/Monster/MonsterSpawnPoint.h"
-
