--- Source/AbyssDiverUnderWorld/Gimmic/SpawnPoint/Ore/OreSpawnPoint.cpp
@@ -1,2 +1,0 @@
-#include "Gimmic/SpawnPoint/Ore/OreSpawnPoint.h"
-
