--- Source/AbyssDiverUnderWorld/Gimmic/SpawnPoint/SpawnPoint.cpp
@@ -1,2 +1,0 @@
-#include "Gimmic/SpawnPoint/SpawnPoint.h"
-
